src/components/register/requirements.txt
pika==1.3.2
python-dotenv==1.0.1
boto3
pillow
psycopg2-binary==2.9.9
SQLAlchemy==2.0.32

